+++ packages/gov-ui-kit/src/components/link/link.tsx
@@ -5,78 +5,86 @@ import { type IconType } from '../icons';
 
 export type LinkProps = React.AnchorHTMLAttributes<HTMLAnchorElement> & {
     disabled?: boolean;
-    /** whether link should open new tab to external location */
+    /** Indicates whether the link should open in a new tab */
     external?: boolean;
     iconRight?: React.FunctionComponentElement<IconType>;
-    iconLeft?: React.FunctionComponentElement<IconType>;
-    /** optional label for the link, defaults to the href if value not provided */
-    label?: string;
-    /** link variants */
-    type?: 'primary' | 'secondary' | 'neutral';
+    /** Label for the link */
+    label: string;
+    /** Optional description */
+    description?: string;
+    /** Variants for link appearance */
+    type?: LinkType;
 };
 
-/** Default link component */
-export const Link: React.FC<LinkProps> = ({
-    disabled = false,
-    external = true,
-    type = 'primary',
-    iconLeft,
-    iconRight,
-    label,
-    href,
-    ...props
-}) => {
-    return (
-        <StyledLink
-            href={disabled ? undefined : href}
-            rel="noopener noreferrer"
-            type={type}
-            disabled={disabled}
-            {...(external ? { target: '_blank' } : {})}
-            {...props}
-            data-testid="link"
-        >
-            {iconLeft && <div>{iconLeft}</div>}
-            <Label>{label ?? href}</Label>
-            {!iconLeft && iconRight && <div>{iconRight}</div>}
-        </StyledLink>
-    );
-};
+export const LINK_VARIANTS = ['primary', 'neutral'] as const;
+type LinkType = (typeof LINK_VARIANTS)[number];
+
+/**
+ * The Link component creates a styled anchor element with optional icon and description.
+ *
+ * @param {LinkProps} props - The properties of the link.
+ */
+export const Link = React.forwardRef<HTMLAnchorElement, LinkProps>(
+    ({ disabled = false, external = true, type = 'primary', iconRight, description, label, href, ...props }, ref) => {
+        return (
+            <StyledLink
+                ref={ref}
+                href={disabled ? undefined : href}
+                rel="noopener noreferrer"
+                type={type}
+                disabled={disabled}
+                {...(external && { target: '_blank' })}
+                {...props}
+                data-testid="link"
+            >
+                <div className="flex gap-x-1 items-center mr-0.5">
+                    <Label>{label}</Label>
+                    {iconRight && <div>{iconRight}</div>}
+                </div>
+                {description && <Description>{description}</Description>}
+            </StyledLink>
+        );
+    },
+);
+
+Link.displayName = 'Link';
 
 type StyledLinkProps = Pick<LinkProps, 'disabled'> & {
     type: NonNullable<LinkProps['type']>;
 };
 
 const StyledLink = styled.a.attrs(({ disabled, type }: StyledLinkProps) => {
-    let className = 'inline-flex items-center space-x-1.5 max-w-full rounded cursor-pointer ';
-
+    let className = 'inline-flex flex-col gap-y-0.25 tablet:gap-y-0.5 max-w-full rounded cursor-pointer ';
     className += variants[type];
-
     className += disabled ? disabledColors[type] : defaultColors[type];
 
     return { className };
 })<StyledLinkProps>`
-    outline: 0; // forcefully setting to remove default Chrome black outline
+    outline: 0; // Remove default Chrome black outline
 `;
 
 const Label = styled.span.attrs({
-    className: 'font-bold truncate',
+    className: 'ft-text-base font-semibold truncate',
+})``;
+
+const Description = styled.p.attrs({
+    className: 'ft-text-sm text-ui-600 truncate',
 })``;
 
 const variants = {
-    primary: 'hover:text-primary-700 active:text-primary-800 focus-visible:ring-2 focus-visible:ring-primary-500 ',
-    secondary: 'hover:text-primary-100 active:text-primary-900 focus-visible:ring-2 focus-visible:ring-ui-0 ',
-    neutral: 'hover:text-primary-700 active:text-primary-800 focus-visible:ring-2 focus-visible:ring-primary-500 ',
+    primary: `hover:text-primary-600 active:text-primary-800 
+        focus-visible:ring focus-visible:ring-primary-200 focus-visible:bg-ui-50 `,
+
+    neutral: `hover:text-ui-800 active:text-ui-800 
+        focus-visible:ring focus-visible:ring-primary-200 focus-visible:bg-ui-50 `,
 };
 
 const disabledColors = {
     primary: 'text-ui-300 pointer-events-none ',
-    secondary: 'text-primary-300 pointer-events-none ',
     neutral: 'text-ui-300 pointer-events-none ',
 };
 
 const defaultColors = {
-    primary: 'text-primary-500 ',
-    secondary: 'text-ui-0 ',
-    neutral: 'text-ui-500 ',
+    primary: 'text-primary-400 ',
+    neutral: 'text-ui-600 ',
 };